+++ relay-server/src/metrics_extraction/transactions/mod.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeMap;
 
 use relay_common::{DurationUnit, EventType, SpanStatus, UnixTimestamp};
-use relay_dynamic_config::{AcceptTransactionNames, TaggingRule, TransactionMetricsConfig};
+use relay_dynamic_config::{TaggingRule, TransactionMetricsConfig};
 use relay_general::protocol::{
     AsPair, Context, ContextInner, Event, TraceContext, TransactionSource,
 };
@@ -20,7 +20,7 @@ use crate::metrics_extraction::utils::{
 };
 use crate::metrics_extraction::IntoMetric;
 use crate::statsd::RelayCounters;
-use crate::utils::SamplingResult;
+use crate::utils::{transaction_source_tag, SamplingResult};
 
 pub mod types;
 
@@ -71,7 +71,11 @@ fn extract_geo_country_code(event: &Event) -> Option<String> {
     None
 }
 
-fn is_low_cardinality(source: &TransactionSource, treat_unknown_as_low_cardinality: bool) -> bool {
+fn is_low_cardinality(source: Option<&TransactionSource>) -> bool {
+    // `None` is used to mark a legacy SDK that does not send the transaction name,
+    // and we assume sends high-cardinality data. See `is_high_cardinality_transaction`.
+    let Some(source) = source else { return false };
+
     match source {
         // For now, we hope that custom transaction names set by users are low-cardinality.
         TransactionSource::Custom => true,
@@ -89,9 +93,9 @@ fn is_low_cardinality(source: &TransactionSource, treat_unknown_as_low_cardinali
         // low-cardinality now.
         TransactionSource::Sanitized => true,
 
-        // "unknown" is the value for old SDKs that do not send a transaction source yet.
-        // Caller decides how to treat this.
-        TransactionSource::Unknown => treat_unknown_as_low_cardinality,
+        // Explicit `Unknown` is used to mark a legacy SDK that does not send the transaction name,
+        // but we assume sends low-cardinality data. See `is_high_cardinality_transaction`.
+        TransactionSource::Unknown => true,
 
         // Any other value would be an SDK bug or users manually configuring the
         // source, assume high-cardinality and drop.
@@ -102,10 +106,7 @@ fn is_low_cardinality(source: &TransactionSource, treat_unknown_as_low_cardinali
 /// Decide whether we want to keep the transaction name.
 /// High-cardinality sources are excluded to protect our metrics infrastructure.
 /// Note that this will produce a discrepancy between metrics and raw transaction data.
-fn get_transaction_name(
-    event: &Event,
-    accept_transaction_names: AcceptTransactionNames,
-) -> Option<String> {
+fn get_transaction_name(event: &Event) -> Option<String> {
     let original_transaction_name = match event.transaction.value() {
         Some(name) => name,
         None => {
@@ -113,15 +114,12 @@ fn get_transaction_name(
         }
     };
 
-    // In client-based mode, handling of "unknown" sources depends on the SDK name.
-    // In strict mode, treat "unknown" as high cardinality.
-    let treat_unknown_as_low_cardinality = matches!(
-        accept_transaction_names,
-        AcceptTransactionNames::ClientBased
-    ) && !store::is_high_cardinality_sdk(event);
+    let source = event
+        .transaction_info
+        .value()
+        .and_then(|info| info.source.value());
 
-    let source = event.transaction_source();
-    let use_original_name = is_low_cardinality(source, treat_unknown_as_low_cardinality);
+    let use_original_name = is_low_cardinality(source);
 
     let name_used;
     let name = if use_original_name {
@@ -130,7 +128,7 @@ fn get_transaction_name(
     } else {
         // Pick a sentinel based on the transaction source:
         match source {
-            TransactionSource::Unknown | TransactionSource::Other(_) => {
+            None | Some(TransactionSource::Other(_)) => {
                 name_used = "none";
                 None
             }
@@ -143,11 +141,7 @@ fn get_transaction_name(
 
     relay_statsd::metric!(
         counter(RelayCounters::MetricsTransactionNameExtracted) += 1,
-        strategy = match &accept_transaction_names {
-            AcceptTransactionNames::ClientBased => "client-based",
-            AcceptTransactionNames::Strict => "strict",
-        },
-        source = &source.to_string(),
+        source = transaction_source_tag(event),
         sdk_name = event
             .client_sdk
             .value()
@@ -172,7 +166,7 @@ fn extract_universal_tags(event: &Event, config: &TransactionMetricsConfig) -> C
     if let Some(environment) = event.environment.as_str() {
         tags.insert(CommonTag::Environment, environment.to_string());
     }
-    if let Some(transaction_name) = get_transaction_name(event, config.accept_transaction_names) {
+    if let Some(transaction_name) = get_transaction_name(event) {
         tags.insert(CommonTag::Transaction, transaction_name);
     }
 
@@ -453,10 +447,11 @@ fn get_measurement_rating(name: &str, value: f64) -> Option<String> {
 mod tests {
     use insta;
     use relay_common::MetricUnit;
-    use relay_dynamic_config::TaggingRule;
+    use relay_dynamic_config::{AcceptTransactionNames, TaggingRule};
     use relay_general::protocol::{Contexts, Timestamp, User};
     use relay_general::store::{
-        self, BreakdownsConfig, LightNormalizationConfig, MeasurementsConfig,
+        self, set_default_transaction_source, BreakdownsConfig, LightNormalizationConfig,
+        MeasurementsConfig,
     };
     use relay_general::types::Annotated;
     use relay_metrics::{DurationUnit, MetricNamespace, MetricValue};
@@ -755,6 +750,7 @@ mod tests {
                 tags: {
                     "measurement_rating": "good",
                     "platform": "other",
+                    "transaction": "<unlabeled transaction>",
                     "transaction.status": "unknown",
                 },
             },
@@ -766,6 +762,7 @@ mod tests {
                 timestamp: UnixTimestamp(1619420400),
                 tags: {
                     "platform": "other",
+                    "transaction": "<unlabeled transaction>",
                     "transaction.status": "unknown",
                 },
             },
@@ -777,6 +774,7 @@ mod tests {
                 timestamp: UnixTimestamp(1619420400),
                 tags: {
                     "platform": "other",
+                    "transaction": "<unlabeled transaction>",
                     "transaction.status": "unknown",
                 },
             },
@@ -788,6 +786,7 @@ mod tests {
                 timestamp: UnixTimestamp(1619420400),
                 tags: {
                     "platform": "other",
+                    "transaction": "<unlabeled transaction>",
                     "transaction.status": "unknown",
                 },
             },
@@ -849,6 +848,7 @@ mod tests {
                 tags: {
                     "measurement_rating": "good",
                     "platform": "other",
+                    "transaction": "<unlabeled transaction>",
                     "transaction.status": "unknown",
                 },
             },
@@ -861,6 +861,7 @@ mod tests {
                 tags: {
                     "measurement_rating": "good",
                     "platform": "other",
+                    "transaction": "<unlabeled transaction>",
                     "transaction.status": "unknown",
                 },
             },
@@ -872,6 +873,7 @@ mod tests {
                 timestamp: UnixTimestamp(1619420400),
                 tags: {
                     "platform": "other",
+                    "transaction": "<unlabeled transaction>",
                     "transaction.status": "unknown",
                 },
             },
@@ -997,53 +999,57 @@ mod tests {
         .unwrap();
 
         insta::assert_debug_snapshot!(metrics, @r###"
-            [
-                Metric {
-                    name: "d:transactions/measurements.a_custom1@none",
-                    value: Distribution(
-                        41.0,
-                    ),
-                    timestamp: UnixTimestamp(1619420402),
-                    tags: {
-                        "platform": "other",
-                        "transaction.status": "unknown",
-                    },
+        [
+            Metric {
+                name: "d:transactions/measurements.a_custom1@none",
+                value: Distribution(
+                    41.0,
+                ),
+                timestamp: UnixTimestamp(1619420402),
+                tags: {
+                    "platform": "other",
+                    "transaction": "foo",
+                    "transaction.status": "unknown",
                 },
-                Metric {
-                    name: "d:transactions/measurements.fcp@millisecond",
-                    value: Distribution(
-                        0.123,
-                    ),
-                    timestamp: UnixTimestamp(1619420402),
-                    tags: {
-                        "measurement_rating": "good",
-                        "platform": "other",
-                        "transaction.status": "unknown",
-                    },
+            },
+            Metric {
+                name: "d:transactions/measurements.fcp@millisecond",
+                value: Distribution(
+                    0.123,
+                ),
+                timestamp: UnixTimestamp(1619420402),
+                tags: {
+                    "measurement_rating": "good",
+                    "platform": "other",
+                    "transaction": "foo",
+                    "transaction.status": "unknown",
                 },
-                Metric {
-                    name: "d:transactions/measurements.g_custom2@second",
-                    value: Distribution(
-                        42.0,
-                    ),
-                    timestamp: UnixTimestamp(1619420402),
-                    tags: {
-                        "platform": "other",
-                        "transaction.status": "unknown",
-                    },
+            },
+            Metric {
+                name: "d:transactions/measurements.g_custom2@second",
+                value: Distribution(
+                    42.0,
+                ),
+                timestamp: UnixTimestamp(1619420402),
+                tags: {
+                    "platform": "other",
+                    "transaction": "foo",
+                    "transaction.status": "unknown",
                 },
-                Metric {
-                    name: "d:transactions/duration@millisecond",
-                    value: Distribution(
-                        2000.0,
-                    ),
-                    timestamp: UnixTimestamp(1619420402),
-                    tags: {
-                        "platform": "other",
-                        "transaction.status": "unknown",
-                    },
+            },
+            Metric {
+                name: "d:transactions/duration@millisecond",
+                value: Distribution(
+                    2000.0,
+                ),
+                timestamp: UnixTimestamp(1619420402),
+                tags: {
+                    "platform": "other",
+                    "transaction": "foo",
+                    "transaction.status": "unknown",
                 },
-            ]
+            },
+        ]
         "###);
     }
 
@@ -1413,12 +1419,15 @@ mod tests {
     }
 
     /// Helper function to check if the transaction name is set correctly
-    fn extract_transaction_name(json: &str, strategy: AcceptTransactionNames) -> Option<String> {
-        let mut config = TransactionMetricsConfig::default();
+    fn extract_transaction_name(json: &str) -> Option<String> {
+        let config = TransactionMetricsConfig::default();
         let aggregator_config = aggregator_config();
 
         let mut event = Annotated::<Event>::from_json(json).unwrap();
-        config.accept_transaction_names = strategy;
+
+        // Logic from `set_default_transaction_source` was previously duplicated in
+        // `extract_transaction_metrics`. Add it here such that tests can remain.
+        set_default_transaction_source(event.value_mut().as_mut().unwrap());
 
         let mut metrics = vec![];
         let mut sampling_metrics = vec![];
@@ -1495,11 +1504,11 @@ mod tests {
     }
 
     #[test]
-    fn test_js_unknown_strict() {
+    fn test_legacy_js_looks_like_url() {
         let json = r#"
         {
             "type": "transaction",
-            "transaction": "foo",
+            "transaction": "foo/",
             "timestamp": "2021-04-26T08:00:00+0100",
             "start_timestamp": "2021-04-26T07:59:01+0100",
             "contexts": {"trace": {}},
@@ -1507,12 +1516,12 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::Strict);
+        let name = extract_transaction_name(json);
         assert!(name.is_none());
     }
 
     #[test]
-    fn test_js_unknown_client_based() {
+    fn test_legacy_js_does_not_look_like_url() {
         let json = r#"
         {
             "type": "transaction",
@@ -1524,8 +1533,8 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
-        assert!(name.is_none());
+        let name = extract_transaction_name(json);
+        assert_eq!(name.as_deref(), Some("foo"));
     }
 
     #[test]
@@ -1542,52 +1551,16 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::Strict);
+        let name = extract_transaction_name(json);
         assert_eq!(name, Some("<< unparameterized >>".to_owned()));
     }
 
     #[test]
-    fn test_js_url_client_based() {
+    fn test_python_404() {
         let json = r#"
         {
             "type": "transaction",
-            "transaction": "foo",
-            "timestamp": "2021-04-26T08:00:00+0100",
-            "start_timestamp": "2021-04-26T07:59:01+0100",
-            "contexts": {"trace": {}},
-            "sdk": {"name": "sentry.javascript.browser"},
-            "transaction_info": {"source": "url"}
-        }
-        "#;
-
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
-        assert_eq!(name, Some("<< unparameterized >>".to_owned()));
-    }
-
-    #[test]
-    fn test_python_404_strict() {
-        let json = r#"
-        {
-            "type": "transaction",
-            "transaction": "foo",
-            "timestamp": "2021-04-26T08:00:00+0100",
-            "start_timestamp": "2021-04-26T07:59:01+0100",
-            "contexts": {"trace": {}},
-            "sdk": {"name": "sentry.python", "integrations":["django"]},
-            "tags": {"http.status": "404"}
-        }
-        "#;
-
-        let name = extract_transaction_name(json, AcceptTransactionNames::Strict);
-        assert!(name.is_none());
-    }
-
-    #[test]
-    fn test_python_404_client_based() {
-        let json = r#"
-        {
-            "type": "transaction",
-            "transaction": "foo",
+            "transaction": "foo/",
             "timestamp": "2021-04-26T08:00:00+0100",
             "start_timestamp": "2021-04-26T07:59:01+0100",
             "contexts": {"trace": {}},
@@ -1596,16 +1569,16 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
+        let name = extract_transaction_name(json);
         assert!(name.is_none());
     }
 
     #[test]
-    fn test_python_200_client_based() {
+    fn test_python_200() {
         let json = r#"
         {
             "type": "transaction",
-            "transaction": "foo",
+            "transaction": "foo/",
             "timestamp": "2021-04-26T08:00:00+0100",
             "start_timestamp": "2021-04-26T07:59:01+0100",
             "contexts": {"trace": {}},
@@ -1614,16 +1587,16 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
-        assert_eq!(name, Some("foo".to_owned()));
+        let name = extract_transaction_name(json);
+        assert_eq!(name, Some("foo/".to_owned()));
     }
 
     #[test]
-    fn test_express_options_strict() {
+    fn test_express_options() {
         let json = r#"
         {
             "type": "transaction",
-            "transaction": "foo",
+            "transaction": "foo/",
             "timestamp": "2021-04-26T08:00:00+0100",
             "start_timestamp": "2021-04-26T07:59:01+0100",
             "contexts": {"trace": {}},
@@ -1632,34 +1605,16 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::Strict);
+        let name = extract_transaction_name(json);
         assert!(name.is_none());
     }
 
     #[test]
-    fn test_express_options_client_based() {
+    fn test_express() {
         let json = r#"
         {
             "type": "transaction",
-            "transaction": "foo",
-            "timestamp": "2021-04-26T08:00:00+0100",
-            "start_timestamp": "2021-04-26T07:59:01+0100",
-            "contexts": {"trace": {}},
-            "sdk": {"name": "sentry.javascript.node", "integrations":["Express"]},
-            "request": {"method": "OPTIONS"}
-        }
-        "#;
-
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
-        assert!(name.is_none());
-    }
-
-    #[test]
-    fn test_express_get_client_based() {
-        let json = r#"
-        {
-            "type": "transaction",
-            "transaction": "foo",
+            "transaction": "foo/",
             "timestamp": "2021-04-26T08:00:00+0100",
             "start_timestamp": "2021-04-26T07:59:01+0100",
             "contexts": {"trace": {}},
@@ -1668,16 +1623,16 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
-        assert_eq!(name, Some("foo".to_owned()));
+        let name = extract_transaction_name(json);
+        assert_eq!(name, Some("foo/".to_owned()));
     }
 
     #[test]
-    fn test_other_client_unknown_strict() {
+    fn test_other_client_unknown() {
         let json = r#"
         {
             "type": "transaction",
-            "transaction": "foo",
+            "transaction": "foo/",
             "timestamp": "2021-04-26T08:00:00+0100",
             "start_timestamp": "2021-04-26T07:59:01+0100",
             "contexts": {"trace": {}},
@@ -1685,29 +1640,12 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::Strict);
-        assert!(name.is_none());
+        let name = extract_transaction_name(json);
+        assert_eq!(name.as_deref(), Some("foo/"));
     }
 
     #[test]
-    fn test_other_client_unknown_client_based() {
-        let json = r#"
-        {
-            "type": "transaction",
-            "transaction": "foo",
-            "timestamp": "2021-04-26T08:00:00+0100",
-            "start_timestamp": "2021-04-26T07:59:01+0100",
-            "contexts": {"trace": {}},
-            "sdk": {"name": "some_client"}
-        }
-        "#;
-
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
-        assert_eq!(name, Some("foo".to_owned()));
-    }
-
-    #[test]
-    fn test_other_client_url_strict() {
+    fn test_other_client_url() {
         let json = r#"
         {
             "type": "transaction",
@@ -1720,30 +1658,12 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::Strict);
+        let name = extract_transaction_name(json);
         assert_eq!(name, Some("<< unparameterized >>".to_owned()));
     }
 
     #[test]
-    fn test_other_client_url_client_based() {
-        let json = r#"
-        {
-            "type": "transaction",
-            "transaction": "foo",
-            "timestamp": "2021-04-26T08:00:00+0100",
-            "start_timestamp": "2021-04-26T07:59:01+0100",
-            "contexts": {"trace": {}},
-            "sdk": {"name": "some_client"},
-            "transaction_info": {"source": "url"}
-        }
-        "#;
-
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
-        assert_eq!(name, Some("<< unparameterized >>".to_owned()));
-    }
-
-    #[test]
-    fn test_any_client_route_strict() {
+    fn test_any_client_route() {
         let json = r#"
         {
             "type": "transaction",
@@ -1756,35 +1676,17 @@ mod tests {
         }
         "#;
 
-        let name = extract_transaction_name(json, AcceptTransactionNames::Strict);
-        assert_eq!(name, Some("foo".to_owned()));
-    }
-
-    #[test]
-    fn test_any_client_route_client_based() {
-        let json = r#"
-        {
-            "type": "transaction",
-            "transaction": "foo",
-            "timestamp": "2021-04-26T08:00:00+0100",
-            "start_timestamp": "2021-04-26T07:59:01+0100",
-            "contexts": {"trace": {}},
-            "sdk": {"name": "some_client"},
-            "transaction_info": {"source": "route"}
-        }
-        "#;
-
-        let name = extract_transaction_name(json, AcceptTransactionNames::ClientBased);
+        let name = extract_transaction_name(json);
         assert_eq!(name, Some("foo".to_owned()));
     }
 
     #[test]
     fn test_parse_transaction_name_strategy() {
-        for (config, expected_strategy) in [
-            (r#"{}"#, AcceptTransactionNames::Strict),
+        for (config_str, expected_strategy) in [
+            (r#"{}"#, AcceptTransactionNames::ClientBased),
             (
                 r#"{"acceptTransactionNames": "unknown-strategy"}"#,
-                AcceptTransactionNames::Strict,
+                AcceptTransactionNames::ClientBased,
             ),
             (
                 r#"{"acceptTransactionNames": "strict"}"#,
@@ -1795,8 +1697,8 @@ mod tests {
                 AcceptTransactionNames::ClientBased,
             ),
         ] {
-            let config: TransactionMetricsConfig = serde_json::from_str(config).unwrap();
-            assert_eq!(config.accept_transaction_names, expected_strategy);
+            let config: TransactionMetricsConfig = serde_json::from_str(config_str).unwrap();
+            assert_eq!(config.deprecated1, expected_strategy, "{}", config_str);
         }
     }